feat: navigate btn
